requirements.txt
streamlit
openai
langchain
langchain-openai
faiss-cpu
python-dotenv
pypdf2
tiktoken
setuptools
